requirements.txt
streamlit
opencv-python
numpy
pillow
phidata
agno
duckduckgo-search
google-generativeai
google-genai
openai
python-dotenv
langchain
langchain-community
langchain-openai
groq
langchain-groq
langchain_google_genai
scikit-image
pytesseract
googlesearch-python
pycountry
pypdf
chromadb